+++ CHANGELOG.md
@@ -1,12 +1,13 @@
 CHANGELOG
 =========
 
-2.2, 2023-09-18
+2.2, 2023-09-19
 ---------------
 - add line number and word-wrap toggles to more controls
 - add diagram export to Tools-menu
 - add option to export diagram with selected entities
 - add refresh-button and copy-button to diagram toolbar
+- add option to copy diagram image to clipboard
 - add YAML formatting options to value editor
 - retain undo history when loading SQL from file to SQL editor
 - fix error on changing column order in trigger/view editor